OC-124 + save authors in DB on edit
