+++ routes/publications/save.js
@@ -1,11 +1,14 @@
+const _ = require('lodash');
 const debug = require('debug');
 
 const api = require('../../lib/api');
 const formHelpers = require('../../lib/form');
+const userHelpers = require('../users/helpers');
 const publishHelpers = require('../publish/steps/helpers');
 
 module.exports = (req, res) => {
   const publicationID = req.params.publicationID;
+  const accessToken = _.get(req, 'session.authOrcid.accessToken');
 
   debug('octopus:ui:debug')(`Saving Publication ${publicationID}`);
 
@@ -30,6 +33,17 @@ module.exports = (req, res) => {
       const updatedPublication = { _id: publicationID, ...publicationData };
 
       debug('octopus:ui:trace')(fields, files);
+
+      if (updatedPublication.collaborators && updatedPublication.collaborators.length > 1) {
+        const { collaborators } = updatedPublication;
+        const orcidIds = collaborators.map((collaborator) => collaborator.userID);
+
+        // Returns data for the collaborators that doesn't exists in our DB
+        const newAuthorsList = await new Promise((resolve) => resolve(userHelpers.checkForNewUsers(orcidIds, accessToken)));
+
+        // Insert New Users List in DB
+        await userHelpers.insertManyUsers(newAuthorsList, res);
+      }
 
       // update publication object
       return api.updatePublication(updatedPublication, (updateErr, updateData) => {
